fix: 🐛 (validate) validator in alert form field in print
